Remove digits array entirely
Now we don't even buffer the digits anymore. That is, we remove
the digits array. We now read the key twice per radix step, once
when counting, and once when moving entries between indices and
output arrays.

This reduces memory and is now on par with zhus on memory.
Instructions did not increase probably because allocation of the
digits array was more costly.

Moreover, we reverse the loop diretions again back to increasing
order, as it was originally. This allows us to use more efficient
runtime iterators array.keys() and array.vals(). This reduces
instructions further and significantly.